bug(#242): regexp
